+++ exporter/exporter.go
@@ -28,6 +28,7 @@ func Start(client rpc.Client) error {
 	go networkLivenessUpdater(client)
 	go eth1DepositsExporter()
 	go genesisDepositsExporter()
+	go checkSubscriptions()
 
 	// wait until the beacon-node is available
 	for {
